Moved some code around for better responsibility divisions
