+++ super_editor/lib/src/core/document.dart
@@ -1,6 +1,5 @@
 import 'package:collection/collection.dart';
 import 'package:flutter/widgets.dart';
-import 'package:super_editor/super_editor.dart';
 
 /// A read-only document with styled text and multimedia elements.
 ///
@@ -55,14 +54,6 @@ abstract class Document {
   /// Returns the [DocumentNode] at the given [position], or [null] if
   /// no such node exists in this [Document].
   DocumentNode? getNode(DocumentPosition position);
-
-  /// Returns a [DocumentRange] that ranges from [position1] to
-  /// [position2], including [position1] and [position2].
-  // TODO: this method is misleading (#48) because if `position1` and
-  //       `position2` are in the same node, they may be returned
-  //       in the wrong order because the document doesn't know
-  //       how to interpret positions within a node.
-  DocumentRange getRangeBetween(DocumentPosition position1, DocumentPosition position2);
 
   /// Returns all [DocumentNode]s from [position1] to [position2], including
   /// the nodes at [position1] and [position2].
@@ -209,69 +200,6 @@ class NodeChangeEvent implements NodeDocumentChange {
 
   @override
   int get hashCode => nodeId.hashCode;
-}
-
-/// A span within a [Document] with one side bounded at [start] and the other
-/// side bounded at [end].
-///
-/// A [DocumentRange] is considered "normalized" if [start] comes before [end].
-/// A [DocumentRange] is NOT "normalized" if [end] comes before [start].
-///
-/// To check if a [DocumentRange] is normalized, call [isNormalized] with
-/// a [Document].
-///
-/// Use [normalize] to create a version of this [DocumentRange] that's guaranteed
-/// to be normalized for the given [Document].
-///
-/// Determining normalization requires a [Document] because a [Document] is the
-/// source of truth for [DocumentNode] content order.
-class DocumentRange {
-  /// Creates a document range between [start] and [end].
-  const DocumentRange({
-    required this.start,
-    required this.end,
-  });
-
-  /// The bounding position of one side of a [DocumentRange].
-  ///
-  /// {@template start_and_end}
-  /// If this [DocumentRange] is normalized then [start] comes before [end], otherwise
-  /// [end] comes before [start].
-  /// {@endtemplate}
-  final DocumentPosition start;
-
-  /// The bounding position of the other side of a [DocumentRange].
-  ///
-  /// {@macro start_and_end}
-  final DocumentPosition end;
-
-  /// Returns `true` if [start] appears at, or before [end], or `false` otherwise.
-  bool isNormalized(Document document) => document.getAffinityForRange(this) == TextAffinity.downstream;
-
-  /// Returns a version of this [DocumentRange] that's normalized.
-  ///
-  /// See [isNormalized] for a definition of normalized.
-  DocumentRange normalize(Document document) {
-    if (isNormalized(document)) {
-      return this;
-    }
-
-    // We're not normalized. To return a normalized version, reverse our bounds.
-    return DocumentRange(start: end, end: start);
-  }
-
-  @override
-  bool operator ==(Object other) =>
-      identical(this, other) ||
-      other is DocumentRange && runtimeType == other.runtimeType && start == other.start && end == other.end;
-
-  @override
-  int get hashCode => start.hashCode ^ end.hashCode;
-
-  @override
-  String toString() {
-    return '[DocumentRange] - start: ($start), end: ($end)';
-  }
 }
 
 /// A logical position within a [Document].
