+++ super_editor/lib/src/default_editor/text_tokenizing/tags.dart
@@ -3,6 +3,7 @@ import 'dart:math';
 import 'package:attributed_text/attributed_text.dart';
 import 'package:characters/characters.dart';
 import 'package:super_editor/src/core/document.dart';
+import 'package:super_editor/src/core/document_selection.dart';
 import 'package:super_editor/src/default_editor/text.dart';
 import 'package:super_editor/src/infrastructure/strings.dart';
